+++ sprocs/auth_admin_assessment_instance.sql
@@ -2,10 +2,10 @@ DROP FUNCTION IF EXISTS auth_admin_assessment_instance (integer, enum_auth_actio
 
 CREATE OR REPLACE FUNCTION
     auth_admin_assessment_instance (
-        assessment_instance_id integer,
+        assessment_instance_id bigint,
         auth_action enum_auth_action,
         auth_data JSONB
-    ) RETURNS TABLE (authorized boolean, auth_user_id integer)
+    ) RETURNS TABLE (authorized boolean, auth_user_id bigint)
 AS $$
 SELECT
     aaci.*
